Move the JSON generation and Kotlin generation to their tooling artifacts
A monolithic CLI module is bad for the classpath, as the tools we want to use depend on different versions of unstable APIs.
