--- redwood-cli/build.gradle
@@ -1,25 +1,0 @@
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-import org.jetbrains.kotlin.gradle.tasks.KotlinJvmCompile
-
-apply plugin: 'java-library'
-apply plugin: 'org.jetbrains.kotlin.jvm'
-apply plugin: 'com.vanniktech.maven.publish'
-apply plugin: 'application'
-
-application {
-  applicationName = 'redwood'
-  mainClass.set('app.cash.redwood.cli.Main')
-}
-
-dependencies {
-  implementation projects.redwoodToolingCodegen
-  implementation projects.redwoodToolingSchema
-  implementation libs.clikt
-}
-
-tasks.named("distTar").configure { task ->
-  task.enabled = false
-}
-tasks.named("assemble").configure { task ->
-  task.dependsOn(tasks.named("installDist"))
-}
