--- redwood-cli/gradle.properties
@@ -1,2 +1,0 @@
-POM_ARTIFACT_ID=redwood-cli
-POM_NAME=Redwood command-line tool
