+++ redwood-tooling-schema/src/main/kotlin/app/cash/redwood/tooling/schema/JsonCommand.kt
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.cli
+package app.cash.redwood.tooling.schema
 
-import app.cash.redwood.tooling.schema.ProtocolSchemaSet
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.arguments.help
